Move `WalkOffCalculationContext` into `ElytraProcess`
